+++ crates/polars-plugin/src/export.rs
@@ -1,0 +1,4 @@
+pub use arrow as polars_arrow;
+pub use polars_core;
+pub use polars_error;
+pub use polars_ffi;
