+++ crates/polars-plugin/src/prelude.rs
@@ -1,0 +1,4 @@
+//! Items commonly needed when writing a plugin expression.
+pub use polars_core::prelude::*;
+
+pub use crate::derive::{CallerContext, DefaultKwargs};
