--- pyo3-polars/pyo3-polars-derive/tests/run.rs
@@ -1,6 +1,0 @@
-#[test]
-fn tests() {
-    let t = trybuild::TestCases::new();
-    t.pass("tests/01.rs");
-    t.pass("tests/02.rs");
-}
