+++ pyo3-polars/pyo3-polars/Cargo.toml
@@ -14,17 +14,14 @@ arrow = { workspace = true, default-features = false }
 libc = "0.2" # pyo3 depends on libc already, so this does not introduce an extra dependence.
 once_cell = "1"
 polars = { workspace = true, default-features = false }
-polars-config = { workspace = true }
 polars-core = { workspace = true, default-features = false, features = ["allow_unused"] }
 polars-error = { workspace = true, features = ["python"] }
 polars-ffi = { workspace = true }
 polars-lazy = { workspace = true, optional = true, features = ["allow_unused"] }
 polars-plan = { workspace = true, optional = true, features = ["allow_unused"] }
+polars-plugin = { workspace = true, optional = true }
 polars-utils = { workspace = true, features = ["serde"], optional = true }
 pyo3 = { workspace = true }
-pyo3-polars-derive = { workspace = true, optional = true }
-serde = { version = "1", optional = true }
-serde-pickle = { version = "1", optional = true }
 thiserror = "2"
 
 [features]
@@ -38,7 +35,7 @@ lazy = [
   "polars-utils",
   "polars-lazy/python",
 ]
-derive = ["pyo3-polars-derive", "polars-plan/python", "polars/include-polars-plan", "serde-pickle", "serde"]
+derive = ["polars-plugin", "polars-plan/python", "polars/include-polars-plan"]
 dtype-full = ["polars/dtype-full", "dtype-decimal", "dtype-array", "dtype-struct", "dtype-categorical"]
 object = ["polars/object"]
 dtype-decimal = ["polars/dtype-decimal"]
