+++ pyo3-polars/pyo3-polars/src/export.rs
@@ -1,5 +1,3 @@
-pub use arrow as polars_arrow;
-pub use polars_core;
-pub use polars_error;
-pub use polars_ffi;
+//! Re-exports used by expression plugins.
 pub use polars_plan;
+pub use polars_plugin::export::*;
